requirements.txt
gradio>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.1
pypdf>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
rich>=13.0.0
